Revisa os types do RenderSubtitle do Teaser
